backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# OpenAI and AI
openai>=1.10.0
tiktoken>=0.5.2

# Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0
unstructured>=0.12.0
python-magic>=0.4.27

# Vector Database
chromadb>=0.4.22
# pinecone-client>=3.0.0  # Uncomment if using Pinecone

# Utilities
numpy>=1.26.3
pandas>=2.1.4
aiofiles>=23.2.1
httpx>=0.26.0

# Monitoring and Logging
loguru>=0.7.2
prometheus-client>=0.19.0

# Testing (optional)
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
